Invalidate cache for .arcconfig
Summary:
Right now, .arcconfig is never invalidated which means that people have to restart Nuclide + shut down the Nuclide server in order to get it updated. This defeats the purpose of having the version checked in the repo.

The way we solve this in other places (.flowconfig, config cache) is to use a LRU cache with 30s timeout.

Reviewed By: hansonw

Differential Revision: D4939411

fbshipit-source-id: 176376f994f78007db92c0e8c050444048800eff