Set `empty_on_fresh_instance: true` on initial Watchman subscriptions to prevent crawl.
Summary:
Previous to this change, we saw requests for `glob(**)` in Eden that were coming
from Watchman. Currently, this is a problem for Eden, as it causes many
Futures to get creating, ultimately resulting in a stack overflow and an
inglorious exit due to a segfault.

According to wez, the `glob(**)` is not due to an explicit Watchman subscription,
but an implicit request for the initial state of the world. This changes
Nuclide's subscriptions to no longer ask for this, so now EdenFS no longer falls
over when Nuclide connects to it.

We also replace a use of `"expression": ["dirname", "relative/path"]` with
`"relative_root": "relative/path"` because according to wez, the latter is more
efficient.

Reviewed By: hansonw

Differential Revision: D4926553

fbshipit-source-id: 7bb808612725edfa48f160306cb52cfc5688f0bf