Add lint rule to ensure that files with @format are pretty printed
Summary:
This copy and pastes the lint rule from www D4883136. I also had to tweak the license header script to allow for `format`.

Note that currently prettier will output trailing commas for calls which is now in the spec but not yet in node. So we can't use prettier for our scripts. www runs babel-node for scripts so we don't have this issue.

Reviewed By: zertosh

Differential Revision: D4933996

fbshipit-source-id: d4f55b1b5a8d1a5d60c570655664ec23610a0553