Change "stop debugging" to "detach debugger" and remove extra stop buttons
Summary:
Our current wording for "Stop Debugging" is ambiguous. In Visual Studio, for example, "Stop debugging" kills the target, whereas "Detach debugger" ends the debug session but leaves the target alone. Our stop also exits the debugger (for exit = debugging pane goes away) for the red X button in the upper right of the pane but not for the stop button on the stepping controls. Overall, this is a bit confusing and un-unified.

This change changes the wording from "stop" to "detach", which is what our actual behavior is. I'm also removing the red X, there are too many ways to close the debugger and this is unnecessary. It's also the only place in Nuclide where we have such a red X. The debugger tab still has its workspace view X to close the tab, which does the same thing as the red X button.

Disable the Stop button when we're not debugging, it doesn't do anything in that case.

Reviewed By: mostafaeweda

Differential Revision: D4950194

fbshipit-source-id: d4e323208b360673adfb994d3f30c0ae06611179