Diagnostics: Support Linter v2 messages
Summary:
I don't think this is necessarily the greatest API ever, but it more or less matches what we already have.
I only added one extension (adding back the "trace" API from the v1 message as an alternative to reference).

https://github.com/steelbrain/linter/blob/master/docs/types/linter-message-v2.md
https://github.com/steelbrain/linter/blob/master/docs/types/standard-linter-v2.md

Note that both v2 and v1 can be handled together since v2 accepts legacy messages.

There are a lot of TODOs here, and those include:
- url & icon handling
- multiple fixes ("solutions"), and the callback variant
- handling the callback-based description

Most of these aren't very useful for us, so we'll add them as they come.

Reviewed By: matthewwithanm

Differential Revision: D4934252

fbshipit-source-id: 746c2a80ef9de44e4fe16e07e3c7490afd9d2a65