Bookshelf: Fix serialization errors and report fatal middlware breakage
Summary:
middleware bteakage should never happen, because in every epic / middlewate, error should already be in place.

In the unlikely case that we encounter a middleware breakage (especially after D4956214 & the logview indicating this only happens when the old `fetchBookmark` fails), we should log it and skip adding it to the `BehaviorSubject`

Changelog: Fix a case where bookshelf serialization gets broken.

Reviewed By: aadisriram, bolinfest

Differential Revision: D4956513

fbshipit-source-id: 72e8a4f49a7a467b4ce49f087e526f3d4220e031