Add simple disconnection logic to Nuclide debugger for mobile clients
Summary:
At the moment, when a device disconnects from the RN/CS packager, Nuclide believes the connection is still alive, and stays in an attached state.

After this diff, the debugger will disconnect and notify the user when:

- The debugger connects and there are no clients attached to the debugger
- The debugger is attached and all clients detach from the packager (i.e. cable disconnection, close emulator)

It does it by checking the number of devices reported by the packager, and running some simple logic to see if there were none, ever, or if the number has changed to 0.

Reviewed By: johnislarry

Differential Revision: D4955896

fbshipit-source-id: 55d8a2cf528e346c942fc25c59964cbe49597895