Add `stale` property for diagnostics
Summary:
This will be useful for a variety of places where updating diagnostics is not
instantaneous. For example, for Flow's push diagnostics, old errors can be
marked as stale while new ones are rolling in. With simpler providers like the
one for `arc lint`, existing errors can be marked as stale as soon as new ones
are requested.

Reviewed By: matthewwithanm

Differential Revision: D4964677

fbshipit-source-id: 69b95d3a420bd8843247ddb45ef67ebd1e649340