Nuclide HHVM debugger does not show fatal exceptions on script exit
Summary:
If a script exits due to an uncaught exception, we just print an error indicating the process status return code returned by PHP. We should be printing the exception with stack trace.

PHP is printing this out to STDERR, but we aren't reading STDERR. Simply forward stderr to the nuclide console.

Reviewed By: johnislarry

Differential Revision: D4965212

fbshipit-source-id: 5b27e88f3d85cefac22760050e51cadd34a57a1e