Install error reporter before activating packages
Summary:
Installing the error reporter during one of the package's activation meant that we could miss errors raised in the activation of earlier packages. By moving it into the UP loader, we'll be sure not to miss any.

Unfortunately, this means a little more console spam for open source, but I think that's okay since it only happens on unhandled exceptions.

This also fixes a bug whereby disabling an re-enabling Nuclide wouldn't install the uncaught exception handler again.

Reviewed By: zertosh

Differential Revision: D4968225

fbshipit-source-id: ef4901ecc0cd356590850b0498cbc711649e5530