Move exit code handling to `getOutputStream()`; defer exit until "close"
Summary:
It's important that the observable returned by `spawn()` complete when
the process exits. However, it's also important that the one returned
by `getOutputStream()` not complete until its streams close (so that we
don't miss any stdout or stderr).

This introduces a wrinkle for us, though, with regard to
ProcessExitErrors: do both utilities have their own handling for
exiting? If so, composing them to create `observeProcess()` would
necessarily be more complicated because `getOutputStream()` would need
to be aware of (and suppress) `spawn()`'s errors.

Instead, I think it makes sense to make `spawn()` a lower-level utility
that is completely unconcerned with any output (including the exit
code). That's the responsibility of `getOutputStream()`.

For the high-level APIs (`runCommand()`, `observeProcess()`, etc.),
there's no API change.

Reviewed By: shushz

Differential Revision: D4975592

fbshipit-source-id: a5c180f777149d3ccc4bb2906829c98af40ea81c