Merge `_makeChildProcess()` into `createProcessStream()`
Summary:
`_makeChildProcess()` is now only called by `createProcessStream()`, so
we might as well merge it in. This allows us to get rid of the type for
its options argument and update `createProcessStream()` to no longer
accept a factory. Since it's now responsible for creating the process,
it no longer has to account for receiving one that's already been
killed.

Making this change made it apparent that we had a potential bug since
we wrote to stdin before adding our listeners, so I fixed that.

Reviewed By: hansonw

Differential Revision: D4976779

fbshipit-source-id: e397b6d7582a5cb7c078686aa6fded2ed111d604