Refactor FileCache part 1
Summary:
The logic and invariants of FileCache are subtle. I'm refactoring the code to draw them out more carefully and explicitly.

This code is a tiny bit less efficient (it removes elements from _buffers one by one rather than removing them all in one go). But it will pay off in terms of code simplification, I promise...!

This first diff in the sequence is really tiny because I want to build up to the overall refactor in small easy-to-read diffs.

Reviewed By: peterhal

Differential Revision: D4854851

fbshipit-source-id: 763ec60be40210965296a35b3538d4d6d655a96d