Refactor FileCache part 4
Summary:
This is an incremental stepping stone towards removing the need for subscribers of FileCache.observeFileEvents(event) from ever needing to call getBufferAtVersion(event) themselves: we will be able to synchronously figure out the buffer associated with the before firing the next(event) signal, and so deliver the buffer to them as part of the observable.

(actually, in future, we won't figure out the *buffer*; we'll just give them the *text content* of the buffer, to stop them messing with the buffer itself).

Anyway, to help write the invariants for this, I want a function that says whether getBufferAtVersion(event.fileVersion) would be able to synchronously figure out the buffer.

Reviewed By: hansonw

Differential Revision: D4854962

fbshipit-source-id: 0e421979a5f06a5c24bf443d278cd805d569580d