Refactor FileCache part 5
Summary:
It used to be that onFileEvent() triggered two things: (1) a call to this._requests.onEvent(), and (2) the downstream subscriber of anyone who wanted to observeFileEvents().

That feels messy. It means that inside the subscriber, we can't be sure that this._requests.onEvent() has yet been called. Therefore we can't be sure that getBufferAtVersion will yet succeed.

I have put the call to this._requests.onEvent() immediately prior to calling _fileEvents.next(). This guarantees that any subscribed to observeFileEvents can know that _requests has been correctly updated.

Reviewed By: hansonw

Differential Revision: D4855005

fbshipit-source-id: 1cbc2b48cfd337d179eed0731d65dadd3b05a30f