Refactor FileCache part 5b
Summary:
Consumers will want to call _tryGetBufferAtVersionSynchronously in the case of FileEditEvent, in order to get the full buffer associated with that edit.

They'll never have any reason to call it after a FileOpenEvent since FileOpen already contains the full buffer.

They'll never have any reason to call it after FileCloseEvent since there's no buffer at that stage.

So I've renamed the function in a fashion that makes it clear who can call it and why. (The invariants in the diff stack leading up to here has proved that this is okay to do).

Reviewed By: hansonw

Differential Revision: D4866324

fbshipit-source-id: 31fccd5e1c888e56568fa446c70a5480a77372ff