Addressed a few potential race conditions
Summary:
Analysis of race conditions is hard! There are three async events to keep track of:

(1) We receive a request for something like AutoComplete or Outline based on a particular version of the file

(2) We receive text edits. But those text edits might lag or be ahead of (1), and they might even skip entirely in case of flakey network.

(3) We perform requests on the LSP server, which might take time, and in the meantime there might be more edits (2).

All of these cases are so hard to test and reason about. I've identified a few of them here and implemented the right behavior.

Reviewed By: hansonw

Differential Revision: D4857091

fbshipit-source-id: 255c7ed02402bac6cd88818fce3a9240f8eac132