sync more carefully when creating textDocumentParams
Summary:
I've eliminated the function "this.createTextDocumentParams(fileVersion)" which was confusing for reason of its side effects - which was to attempt to get the nuclide server and the LSP server synced to that particular version.

There's usually no need for the former. I think the code is more clear if the callers explicitly indicate that they're syncing the LSP server, and why:

The general contract is, "For every LanguageService operation, the service should await until LSP is at the exact right version (and fail if it can't get there)".

There are a couple of exceptions... for formatSource we do actually need NuclideServer to also be at the exact same version; for TypeCoverage we don't even specify the version of the file.

Reviewed By: peterhal

Differential Revision: D4857098

fbshipit-source-id: 1f056ef492b917433711362c8dbc1a2ea24fa0ec