Add a Restart Debugger command to Nuclide debugger
Summary: This adds a "restart debugger" command. It is opt-in, debuggers must indicate they support it for it to be available. If enabled for the current debugger, a new "restart debugger" button will appear to the left of the debugger stepping controls. Restarting the debugger stops the current debug session, and starts a new one with the exact same launch/attach settings. This lets users quickly reload and restart the session after making code changes.

Reviewed By: yinghuitan

Differential Revision: D4997352

fbshipit-source-id: 8295f40ff0133a5b3d7cf617e948a671bfa40bf9