Add Imperative Console Service API
Summary:
This is something I've been wanting to do for a looooong time. It exposes a familiar console API and is Rx-free. (On either side of the service boundary, we'll probably use Rx still...I just don't want to require it for interacting with the service.)

Right now, there's a lot of redundancy because it's supporting both APIs. Once we move things over from the "output-provider" service, there's a lot of cleanup that could happen.

Reviewed By: stepanhruda

Differential Revision: D5000347

fbshipit-source-id: da6035d6329e855da0972d9c725868645a4914ce