Fix no-cross-atom-imports when run from outside the Nuclide dir
Summary:
This fixes an issue where `no-cross-atom-imports` would throw if ran on a file that required a node builtin module (like `fs`) and the cwd was outside of the Nuclide dir.

`resolveFrom` returns the module name when resolving builtins - instead of `null`. So when it came time to locating the nearest `package.json` for a builtin, `path.resolve` would resolve the module path from the cwd. When run from within the Nuclide dir, you got the base `package.json`, but when run from outside, it failed.

Reviewed By: vjeux

Differential Revision: D5018834

fbshipit-source-id: a96965c4d01dc24dcc9c1fc2fffa6fc0e7d8f4f1