Make mobile JS hot reload interoperate with the debugger
Summary:
This diff fixes two problems seen after hot reloading code for the ComponentScript sample app.

The first one, in CompilerScriptMapping, allows for the SourceMap values to be replaced after a reload.

The second one, accross the other files, makes it so breakpoints are re-sent after a new bundle has been reloaded and parsed. This fixes a sequence problem where breakpoints would not resolve correctly in FileCache#getUrlFromFilePath because the _nuclidePathToFileData and filePath would contain different values.

Reviewed By: yinghuitan

Differential Revision: D5012011

fbshipit-source-id: 6539646f83054112b52b3bda4e99cc133e191f05