Add replay to the poller
Summary:
The poller broadcasts every 3 seconds the list of connected devices. Even if there was already a broadcast with devices, a late subscriber would have to wait for the next broadcast to get the list of devices.
With this replay, a late subscriber will get the last list that was broadcasted and from that point on will start receiving up-to-date lists.

Reviewed By: stepanhruda

Differential Revision: D5031570

fbshipit-source-id: bdce642589fcea008dd61320493efa733ed3fc47