Split the device panel into two panels
Summary:
This makes the first panel show information of the available devices types and hosts. In the future it will show actions like 'start an emulator'
Upon selecting a device on the first panel, the second panel appears with information regarding only the selected device along with the available actions.
This is not finished, in the coming diffs i'll finish this.

Reviewed By: stepanhruda

Differential Revision: D5024212

fbshipit-source-id: 9e19f29d25ca20efd6fc953ca5e09124972e783a