Limit pane search to activate pane's container in nuclide-move-item-to-available-pane
Summary:
Until Atom 1.17, there was only one pane container in a normal Atom
configuration: the one that housed the text editors in the center of
the workspace. As of 1.17, however, there are multiple pane containers
--one for the center, and one for each dock--and all of their panes are
returned by `atom.workspace.getPanes()`. That means that this utility
could end up moving an item to a new container. This change addresses
that by limiting the search to the container of the provided pane.

Reviewed By: mostafaeweda

Differential Revision: D5041403

fbshipit-source-id: 355ca8492ba7e3955ff4baf84c94ca69b48bf1ae