Fix unused flow supression and make code flow happy
Summary: `AttachProcessInfo` had two instances of requiring `fb/services`. While one had a `// $FlowFB`, the other did not, and Flow complained that the case where it actually existed was an unused suppression, while on the other callsite it wanted a suppression. Raised a task internally with the Flow team; landing this to unblock CircleCI.

Reviewed By: jgebhardt

Differential Revision: D5044075

fbshipit-source-id: ecd87dc0c4f9c798eef9e45ee6761ba5e87203f0