Some improvements on the process table
Summary:
Ranemed some functions to processes instead of processTable, because those are in fact arrays and not maps.
Make the pid a number.
Show all the pids, which is now possible because we have a filter.
Remove the title and make the filter input box more descriptive as a replacement.
Show the pids in descending order so that more recent processes are shown first.

Reviewed By: stepanhruda

Differential Revision: D5039138

fbshipit-source-id: 03990e74737873686c6c348962b33774e221f1d1