Unregister debug events from new channel
Summary: Chrome protocol parser has code to do validation and generate error/warning messages for unimplemented protocols. It used to show a bunch of warning/errors in dev console of chrome process. Now we are hosting it in Nuclide, we will see these error/warning messages in Nuclide dev console. Functionality wise, there is no problem, but in order to not confuse user, let's unregister event handling for new channel for now.

Reviewed By: mostafaeweda

Differential Revision: D5058849

fbshipit-source-id: acad399a84770ede729b04281c26a826497f8172