Make debugger panes each their own workspace view item - Layout debugger - 4/N
Summary:
This change explicitly lays out the debugger panes in the correct locations on launching using the workspace APIs for Atom version >= 1.17. The workspace.open API cannot reliably lay out the panes the way we need it to so unfortunately so we explicitly show and hide them in specific docks.

For downlevel compatiblity, I am leaving the existing workspace.open()-based implementation in place. This can be removed once we require >= 1.17 for Nuclide.

Reviewed By: yinghuitan

Differential Revision: D5026539

fbshipit-source-id: 1665181b3e6628b34d74d7ef57eb65d2739060c8