Make debugger panes each their own workspace view item - remember pane locations - 7/N
Summary:
This change remmebers which dock each debugger pane was last positioned in between debugging sessions, and accross restarting Nuclide. Before hiding panes, we track which dock they are in, and serialize that state to local storage. This is restored when showing the debugger again.

Panes will have an affinity to the dock they were last positioned in, and will maintain relative order with respect to other panes owned by the debugger. No attempt is made to spawn them into the exact same tab configuration within a dock - the Atom APIs don't offer me any way to examine rows and cols of the various splits, so I can't recreate the layout precisely.

A new "Reset Layout" command is added to forget all this state and return the debugger to its default layout.

Remembering panes the user has explicitly hidden is not implemented yet in this diff. Any hidden panes will spawn back to their last visible location every time the debugger is started.

Reviewed By: yinghuitan

Differential Revision: D5026556

fbshipit-source-id: d28279a4cd36ca9907269a36bd11925ee81f4704