Provide HostServices to languages
Summary:
Atom can already call into LanguageService, e.g. to compute "go-to-definition" for a given caret location.

This diff adds the reverse direction, where the LanguageService can call into Atom.

We could have done some simple cases in the forward direction, by having Atom call a method like "observeConsoleMessages" or "observeDiagnostics" and then have the LanguageService send each next message. But...

(1) When the language starts up, it might already be producing diagnostics even before observeDiagnostics has been called. We wouldn't want them to be lost. (This is a theoretical race condition because no one really produces diagnostics very fast. But it will soon become a real race condition when the langauge service wants to start having Atom display its status).

We could solve the race condition by delaying the language service from starting up until such time as observeDiagnostics gets called. But that's needlessly synchronous, and a bit fussy, and will block if the language config doesn't end up calling observeDiagnostics. We could buffer things up until such time has observeDiagnostics gets called. But buffers aren't nice, and this one could grow unbounded if the language config doesn't end up calling observeDiagnostics.

(2) In the case of dialogs, the language service needs to know when the user has dismissed the dialog, and what the result was. We can't do this just with observables.

(3) Notionally, once you add an API in Atom that languages can call into, it should be as simple as just adding an API which languages can chose to call or not, as they wish. But if you do it by adding a language service method like "observeConsoleMessages", then every single language service is affected, and has to return an observable, even if it's just a dummy empty observable or a throw "not yet implemented" message.

There's one thing to note which might be obvious but I want to spell it out. If there are two windows of Nuclide spun up, they will each communicate to the same server instance. So we can't have a single server-side global variable called 'host' which the language server uses to communciate back to Atom. Instead we have to pass down the correct instance of 'host' to each LanguageService

Reviewed By: hansonw

Differential Revision: D4999106

fbshipit-source-id: 8fd44122eb8780cad50ad5bcaee9a269003d4f18