Add dialog support to HostServices
Summary:
This diff allows a language service to cause a notification to pop up, with optional buttons, and see which button (if any) the user clicked.

I could have made the dialog methods return a Promise rather than a ConnectedObservable. After all, the dialog only ever returns a single value. But nuclide-rpc times promises out after 60s, so that wouldn't have worked. (users can and will leave dialogs open for longer).

I made it so the dialog appears only when you connect to the ConnectedObservable. And when you unsubscribe, then the dialog gets dismissed (if it's still visible).

Also, thanks to HostServices aggregation, we have the opportunity to close all outstanding dialogs for a given language service (i.e. for a given fork of HostServices) if that fork gets disposed of.

Reviewed By: hansonw

Differential Revision: D4999234

fbshipit-source-id: cc822e8e9b698c38e0544e9512fd79bcff450886