eliminate LspProcess class
Summary:
This diff is in preparation for managing the LSP lifecycle.

I'm eliminating the class LspProcess. It was a bit of a weird object - it didn't encapsulate anything (since LspLanguageService kept peeking inside its private fields), it basically just did some helper methods. Those helper methods need to be popped out in preparation for a proper solution of LSP lifecycle (restart etc.)

VSCode also doesn't have an analog for LspProcess.

Reviewed By: hansonw

Differential Revision: D5004702

fbshipit-source-id: 44be8514335f74b1d4dbf744271d0ac341fd1618