Support to avoid observeDiagnostics race
Summary:
It's possible that we spawn an LSP process, and it starts generating diagnostics, but we only observeDiagnostics on it afterwards. If so, then diagnostics would be lost.

This is just a theoretical race for now, since we observe pretty quickly, and LSP servers seem to take a while before they start reporting diagnostics. But it's still a race, and it will get worse once we start observing diagnosticStatus (which will likely be reported instantly).

One fix would be to just buffer the LSP's diagnostics until such time as observeDiagnostics is called. This is possible, but I don't like unbounded buffers!

The real fix, I believe, is to move 'reportFileDiagnostics' into a method in HostServices, not an observable at all. That's because HostServices will already be available, and it has means to aggregate and invalidate on close, so we wouldn't need to rewrite the "aggregate+invalidate" code everywhere (in MultiProjectLanguageService, in FlowService, in LspLanguageService).

But right now I'm making a smaller fix. I'm merely making our language service factory await until someone has observed diagnostics. This is more synchronous than ideal, but probably not really an issue in practice.

Reviewed By: peterhal

Differential Revision: D5027390

fbshipit-source-id: 8a0f6cead646fe13be520a8db09867c9d222515c