lifecycle of process, restart, dispose
Summary:
LSP lifecycle management. Including automatic restart as needed in response to a server crash. Also includes error logging. And guards against re-entrancy.

This is a HUGE diff! Sorry about that. I honestly didn't see any way to approach it incrementally (beyond the two refactoring diffs previously in the stack). That's because lifecycle needed a comprehensive rewrite, and lifecycle is a cross-cutting concern that touches everything.

I tried to keep it as close as possible to how VSCode works. That's because I figured we'd be safest this way -- if LSP servers make assumptions about the way VSCode works, we'll likely work the same way. The lifecycle flow chart is here:
https://github.com/Microsoft/language-server-protocol/issues/227

(That link also has an  explanation of how we avoid doing automatic restart in case another Nuclide window has taken over the Hack persistent connection).

Note: there was an interesting architectural design question of how to implement 'observeDiagnostics'. Hack implemented it one way, by catching connection exit, and then removing it from the MultiProjectLangaugeService._processes analog, and then re-adding it.

Flow implemented it another way, with an observable of connections. It wrapped this with some subtle logic, similar to ensureInvalidations.
diffusion/FBS/browse/master/fbobjc/Tools/Nuclide/pkg/nuclide-flow-rpc/lib/FlowSingleProjectLanguageService.js;7836ebfc80cd9103eedd2c06ab9bbcf9fd2a312c$287

I picked a third way, where the LspLanguageService lives on persistently and is able to restart itself (for similarity with the VSCode flowchart). And I used

  this._diagnosticUpdates: Observable<Observable<Diagnostics>>

  const perConnectionUpdates = new Subject();
  this._diagnosticUpdates.next(perConnectionUpdates);

  this._lspConnection.onDiagnosticsnotification(perConnectionUpdates.next);

Using subjects makes it easy to see that I'm only calling subscribing to the LSP onDiagnosticNotification exactly once per connection, and feels a bit closer in spirit to how VSCode did it. It also makes for a really elegant implementation of observeDiagnostics:

  return this._diagnosticUpdates
    .mergeMap(perConnectionUpdates =>
      ensureInvalidations(this._logger, perConnectionUpdates.map(convert)),
    )
    .publish();

There are some other noteworthy parts to the diagnostics story. I worked hard to ensure that any diagnostics sent eagerly from the LSP server will for sure be received by Nuclide. That's complicated by the fact that AtomLangaugeServices might invoke the observeDiagnostics method later. I put notes in the code to indicate how this is achieved.

Reviewed By: hansonw

Differential Revision: D5014975

fbshipit-source-id: 821cc029bd05bbdef2b965e4ab94247eebbb8868