Catch git commands when fetching files from repo
Summary:
When fetching files using repo, we first get the list of git projects and then we run ls-files on all of them.
At this point, if one git command fails, it all would fail and then nuclide would report that no files exist at all.
This catches this scenario.

Reviewed By: sas

Differential Revision: D5092279

fbshipit-source-id: 990fde0713e7b8208837fd4bd08266aff3070b5b