Always show "uncommitted/head/stack changes" section
Summary:
It used to be that the "uncommitted changes" section of the filetree would automatically disappear if there were no uncommitted changes.

But now that it has a drop-down to select between "uncommitted/head/stack", it's more useful to have this section always present.

(I figure we might as well make it always-present for everyone, rather than complicating the matter and having it always-present only for the people who pass the gatekeeper for the new uncommitted/head/stack feature.)

Something I'm uneasy about:
https://pxl.cl/7pqj

Do we want it to display the words "No changes"? Maybe we do because the section header looks a bit weird with nothing under it. Maybe we don't because the words "No changes" look wordy.

https://pxl.cl/7ps0

What I did instead was to display the words "No changes", but indent them a bit so they're the same indentation as files are. This felt a bit nicer than left-aligned: it looked less like a section heading.

(Note: there was a hidden bug before where the MultiRootChangedFilesView failed to render those words in the case where there were one or more project roots but none of them had changes. We never observed that before because the entire section was hidden).

Reviewed By: mostafaeweda

Differential Revision: D5092346

fbshipit-source-id: 205dd454db110e7e342761c691ca5ac4203124d4