Make proper Atom service out of Java debugger client package
Summary:
This removes the need for the cross atom imports and the try/catch/require nonsense when importing the Java debugger from the Buck DeployEventStream class.

Cleaned up the constructor for AdbProcessInfo - the parameters were getting out of hand. It now takes in an object as params so they can be named and easier to read at the call sites.

Reviewed By: yinghuitan

Differential Revision: D5096267

fbshipit-source-id: de7d8af5b8d9b38afca5e031d23df236c646fb63