Add Keymap for triggering datatips via the keyboard
Summary:
This adds an api to trigger datatips located under the editor cursor. With the addition of modifier datatyps, this means that before the hyperclick keychord (cmd-alt-enter) is completed with cmd-alt, the user receives a preview of where they will jump with the keyboard.

Additionally, users can simple hold `alt` to get a typehint of the symbol below whereever the cursor is posiitioned.

Reviewed By: vjeux

Differential Revision: D5056544

fbshipit-source-id: 317db413497c7147ee5e18ec392fbb13f71deb56