nuclide-move-item-* commands should use current pane
Summary: Previously, this would always use the active pane (even if the command was dispatched on another, as was the case in Diff View). This update makes it try to use the Pane of the element on which the command was dispatched, only falling back to the active pane if it can't find one.

Reviewed By: aadisriram

Differential Revision: D5106182

fbshipit-source-id: 617091332ee84a5061b8e7afb79a5afb466a1bab