Add a service that can open SSH tunnels
Summary:
I have run into multiple cases where not being able to contact localhost from a remote directory is a limiting factor. A good example of this is if you want to run lldb remotely (because the source and all tooling is on the remote machine) across a TCP connection, and the debugged binary is running locally (or on a physical device connected to localhost!). We would need to add protocol multiplexing to nuclide-server, which seems like significant effort. This is a cheap and transparent solution in the meantime, although it comes with the nuisance of needing to deal with 2FA when opening a tunnel.

Note: This will only ever be opening reverse tunnels from a remote machine you're connected to to localhost. The tunnels never go the other way.

Reviewed By: matthewwithanm

Differential Revision: D5097128

fbshipit-source-id: 00623aad4f633c8db301aa51bf175def95f97c0b