Create a provider-based busy signal API
Summary:
For `atom-ide-ui`, the consumer-based API is problematic - we don't want to force people to import a library to use the busy signal!

This creates a provider-based API for the busy signal that instantiates a "deduped busy signal provider" and gives it to the consumer, exposing:

- `reportBusy`
- `displayMessage`
- `dispose`

Additionally, this is named `atom-ide-busy-signal` as this will be part of the `atom-ide-ui` distribution.

All the use cases in Nuclide use deduping so I think it makes sense to only expose the deduped provider.

The following stack of diffs converts all the providers into consumers, and I'll clean this package up at the very end.

Reviewed By: matthewwithanm

Differential Revision: D5100323

fbshipit-source-id: 34ff2df8e9ec7954dcfec6880246d0a53ba74fd9