Remove source map caching
Summary:
Here's the scenario this diff fixes:

1. Attach JS debugger to device
2. Make NO local edits to your JS code
3. Reload the JS from the packager
4. Breakpoints won't hit!

However, they will hit if you make local edits and reload.  This is because the chrome devtools does some caching using the base64-converted source map, and in the scenario above, the SM doesn't change, so it gives our setBreakpointByUrl messages a cached url, which we can't properly translate.

I think the best thing to do is just get rid of this caching since it doesn't give us a performance benefit, and any alternative workarounds I can think of are way more complicated.

Reviewed By: ebluestein

Differential Revision: D5115554

fbshipit-source-id: d2d6ed24b851fad6b5106a9be68106400af31b5f