Remove `getLogger` from nuclide-logging and call log4js directly
Summary:
This removes `getLogger` from nuclide-logging, which made getting loggers lazy,
something that is mostly accomplished by our babel inline requires transform. This drastically
simplifies nuclide-logging to simply become a way to load configuration objects which can
be set as the log4js config in nuclide's entry point.

This codemods every module that used `getLogger` in nuclide-logging to require it from log4js directly.
The codemod also infers the package name and uses it as the logging category, leading to more informative
logs since everything used to use a category of 'nuclide'.

Reviewed By: hansonw

Differential Revision: D5107477

fbshipit-source-id: 79531218a61c049f3783bb08d8f55677f0da723d