*easy* Don't log RequestCancelled errors
Summary:
VSCode reports every error other than RequestCancelled in its console. That's because RequestCancelled is part of the normal operation of an LSP server.

We we only reporting errors in the nuclide log, not in the user-facing console. But I think that even the nuclide log shouldn't report RequestCancelled because that would be too noisy.

Reviewed By: arxanas

Differential Revision: D5110487

fbshipit-source-id: 7bd6ad342197fcf7f340abe07348212596482526