*easy* if stderr is printed while connecting, display it as info
Summary:
In the command-line world, we usually use stderr not for *errors* but for informational out-of-band messages. I think that likewise, if Nuclide is connecting to an LSP server and it prints some stuff to stderr but succeeds in responding to the "initialize" request, then it was presumably using stderr for informational rather than error-reporting purposes.

This change makes stderr output in this case be shown as 'info' rather than 'error' in the logs. It also bumps up the character count to 600, since the previous limit of 300 was feeling a bit cramped.

Reviewed By: arxanas

Differential Revision: D5110493

fbshipit-source-id: c4da519e9eacfab359525bebc91eb2deee42819e