Format as you type
Summary: The language server protocol supports a notion of "on-type formatting", in which the language server formats your code for you as you type. It notifies the client about a set of trigger characters (such as `}` and `;`) which cause an `onTypeFormatting` request to be sent. The server then sends edits back to apply to the buffer.

Reviewed By: ljw1004

Differential Revision: D5066723

fbshipit-source-id: 3e032f1f9e9a6631e55da24eeffbfc93fb33ff50