Make most elements unselectable by default
Summary:
This adds a top-level rule limiting selection on most elements, cascading
down from the react root element for nuclide components. Unlike the web, where this
is expected and is more natural, having selectable text for ui bits feels really
unnatural in native apps.

[Most Atom components](https://github.com/atom/tree-view/blob/66aca880ced946874ce9629efe0ec46b6bb17ddd/styles/tree-view.less#L11)
[do this by default](https://github.com/atom/status-bar/blob/a65b16e12b18f1b4d8e33857c6b3739e40b4b17b/styles/status-bar.less#L10-L11)
and this is easily opted out of by setting `user-select: text`. Moreover, by setting this
on the top level only, inputs, textarea, and others continue to use text cursors and have selectable content
without explicitly opting out of it.

An alternative approach to this would be to add the cursor and user-select lines to the selector for the
parent element of each nuclide package with ui. It would make this a bit more explicit at the cost of
needing this for every package. This might be a good option since it's pretty weird to style the react
mount root (this was the only selector available at a top level commonly across packages).

Reviewed By: jgebhardt

Differential Revision: D5110107

fbshipit-source-id: f40a803fc6eebfbdbac4d074820fe4eb671b88c5