Move the workspace-views 1.17 compatibility shim into nuclide-commons-atom
Summary:
`atom-ide-ui` packages can't have a formal dependency on `nuclide-workspace-views`, since it doesn't exist in the standalone distribution.

It'll probably another ~2 weeks before we can totally kill 1.16 support in Nuclide, so to support interop we need a compatibility shim that hooks into the service hub at runtime.

I created `nuclide-commons-atom/workspace-views-compat` to facilitate this. `atom-ide-ui` packages can reach in and call `consumeWorkspaceViewsCompat`, which uses Docks in 1.17+, but otherwise dynamically consumes the `nuclide-workspace-views` service.

`atom-ide-ui` is going to be 1.17+ only so the second case is only going to happen inside of Nuclide.

Reviewed By: matthewwithanm

Differential Revision: D5125625

fbshipit-source-id: e43f499b17f473fe904f33fb33097401f6e44454